requirements.txt
click==8.1.3
colorama==0.4.5
colormath==3.0.0
libscrc==1.8.1
networkx==2.8.6
numpy==1.23.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
pyqtgraph==0.12.4
pyserial==3.5
python-dotenv==0.20.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
spectra==0.0.11
